nginx: make default server conf a site conf
The default servers for HTTP and HTTPS are currently defined in a
`conf.d/` file. This commit installs the file as a site in the
`sites-available/` directory that is symlinked in `sites-enabled/`. The
conf file is also named the root domain name since that will serve as
the default server.