nginx: use root site for Lets Encrypt challenges
There was no way for cert renewals to succeed using the existing site
configs. The only config that allowed non-HTTPS requests to
`/.well-known/acme-challenge` was the temporary one used for fetching
the initial cert. This commit adds a location block for the challenge
path to the root site conf and installs it earlier in the role so that
it can be used to get the initial cert.