nginx: use the default nginx.conf
It is suggested that the `nginx.conf` that comes with Nginx is used
since an update to the package could overwrite the file. Now that
customizations to the configuration have been moved into files that are
installed in the `conf.d/` directory, a custom `nginx.conf` is no longer
needed.