primary-user: user creation is idempotent
The password is provided to ansible in plaintext and piped through the
`password_hash` filter. This filter generates a random salt if one is
not provided resulting in a different hash each time. This makes the
task not idempotent. Add the `update_password: on_create` arg to the
task to ensure the password hash is only set when the account is
created, thus making it idempotent.