custom-motd-art: use python for multibyte handling
The MOTD script needs to determine the lengths of lines in order to
center the neofetch output. The current implementation uses `awk`, which
seemed to handle multibyte characters correctly. However, I discovered
that older versions -- such as that of the pre-installed `awk` on Ubuntu
Server -- do not handle multibyte chars correctly when calculating the
line lengths.

Use Python 3, which (hopefully) reliably handles multibyte chars
correctly.