fix(nginx): move remote temp out of ansible home
The ansible home dir is created by sudo with restrictive permissions.
When a task runs as an unprivileged user it runs into permission errors
creating a task temp dir (not 100% certain of this, but that is what
appears to be happending). Move the ansible remote temp dir out of the
ansible home dir to prevent these permission errors.