nginx-nextcloud-server: improve idempotence
Reindex NextCloud files task used to be a handler but was moved. No
`when` was added to the action, dropping its idempotence. Add a `when`
so that it is only run when changes to files are made. In this case, if
any of the demo files are deleted in the previous task.